# docs/swagger.yaml
...
swagger: "2.0"
info:
  description: "API of Metadata Service for LAAWS"
  version: "1.0.0"
  title: "LAAWS Metadata Service API"
  termsOfService: "http://www.lockss.org/terms/"
  contact:
    name: "LOCKSS Support"
    url: "http://www.lockss.org/"
    email: "support@lockss.org"
  license:
    name: "Modified BSD License"
    url: "http://www.lockss.org/support/open-source-license/"
host: "laaws.lockss.org:443"
basePath: "/"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- basicAuth: []
paths:
...
  /urls/openurl:
    get:
      tags:
      - "urls"
      summary: "Performs an OpenURL query"
      description: "Provides the URL that results from performing an OpenURL\
        \ query. With query parameters inline"
      operationId: "getUrlsOpenUrl"
      produces:
      - "application/json"
      parameters:
      - name: "params"
        in: "query"
        description: "The OpenURL parameters"
        required: true
        type: "array"
        items:
          type: "string"
      responses:
        200:
          description: "The data related to the performed OpenURL query"
          schema:
            $ref: "#/definitions/urlInfo"
        401:
          description: "Unauthorized request"
        500:
          description: "Internal server error"
        503:
          description: "Some or all of the system is not available"
  /urls/doi/{doi}:
    get:
      tags:
      - "urls"
      summary: "Gets the URL for a DOI"
      description: "Provides the URL for a DOI given the DOI"
      operationId: "getUrlsDoi"
      produces:
      - "application/json"
      parameters:
      - name: "doi"
        in: "path"
        required: true
        type: "string"
        description: "The DOI for which the URL is requested"
      responses:
        200:
          description: "The URL for the specified DOI"
          schema:
            $ref: "#/definitions/urlInfo"
        401:
          description: "Unauthorized request"
        500:
          description: "Internal server error"
        503:
          description: "Some or all of the system is not available"
...
securityDefinitions:
  basicAuth:
    description: "HTTP Basic Authentication. Works over `HTTP` and `HTTPS`"
    type: "basic"
definitions:
...
  urlInfo:
    type: "object"
    required:
    - "params"
    - "urls"
    properties:
      params:
        type: "object"
        description: "The parameters that define the resulting URLs"
        additionalProperties:
          type: "string"
          description: "The value of the parameter for the parameter name"
      urls:
        type: "array"
        description: "The URLs"
        items:
          type: "string"
    description: "The information related to a resulting list of URLs"
...